:recycle: Extract xxx_config.go
