+++ config-simple.json
@@ -1,0 +1,8 @@
+{
+  "job":{
+    "subscription":"projects/scenic-doodad-617/subscriptions/akm-pipeline01-job-subscription"
+  },
+  "progress":{
+    "topic":"projects/scenic-doodad-617/topics/akm-pipeline01-progress-topic"
+  },
+}
