+++ download_config.go
@@ -1,0 +1,5 @@
+package main
+
+type DownloadConfig struct {
+	Worker *WorkerConfig `json:"worker,omitempty"`
+}
