+++ examples/ruby/tmp/workspace/downloads/bucket1/dir1/test.txt
@@ -1,0 +1,1 @@
+test1
