+++ examples/ruby/tmp/workspace/uploads/bucket1/dir1/test-test.txt
@@ -1,0 +1,1 @@
+test1
