+++ job.go
@@ -20,16 +20,6 @@ import (
 	logrus "github.com/sirupsen/logrus"
 )
 
-type CommandConfig struct {
-	Template []string            `json:"-"`
-	Options  map[string][]string `json:"options,omitempty"`
-	Dryrun   bool                `json:"dryrun,omitempty"`
-}
-
-func (c *CommandConfig) setup() *ConfigError {
-	return nil
-}
-
 type Job struct {
 	config *CommandConfig
