+++ progress_notification.go
@@ -12,19 +12,6 @@ import (
 	logrus "github.com/sirupsen/logrus"
 )
 
-type ProgressNotificationConfig struct {
-	Topic      string            `json:"topic"`
-	LogLevel   string            `json:"log_level"`
-	Attributes map[string]string `json:"attributes,omitempty"`
-}
-
-func (c *ProgressNotificationConfig) setup() *ConfigError {
-	if c.LogLevel == "" {
-		c.LogLevel = logrus.InfoLevel.String()
-	}
-	return nil
-}
-
 type ProgressNotification struct {
 	config    *ProgressNotificationConfig
 	publisher Publisher
