+++ upload_config.go
@@ -1,0 +1,6 @@
+package main
+
+type UploadConfig struct {
+	Worker           *WorkerConfig `json:"worker,omitempty"`
+	ContentTypeByExt bool          `json:"content_type_by_ext,omitempty"`
+}
